+++ pants-plugins/project_version/BUILD
