+++ pants-plugins/project_version/register.py
@@ -1,0 +1,10 @@
+from project_version.targets import ProjectVersionTarget
+import project_version.custom_rules as custom_rules
+
+
+def target_types():
+    return [ProjectVersionTarget]
+
+
+def rules():
+    return [*custom_rules.rules()]
