+++ pants.toml
@@ -8,8 +8,8 @@ backend_packages = [
   "pants.backend.python.lint.black",
   "pants.backend.python.lint.flake8",
   "pants.backend.python.typecheck.mypy",
-
-  "experiment",
+  # plugins
+  "project_version",
 ]
 
 [python]
